app/src/androidTest/java/com/simplu/pingpongscore/MatchDatabaseTest.kt
package com.simplu.pingpongscore

import androidx.test.runner.AndroidJUnit4

/*
 * Copyright 2019, The Android Open Source Project
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import androidx.room.Room
import androidx.test.platform.app.InstrumentationRegistry
import com.simplu.pingpongscore.database.Match
import com.simplu.pingpongscore.database.MatchDatabase
import com.simplu.pingpongscore.database.MatchDatabaseDao
import org.junit.Assert.assertEquals
import org.junit.After
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith
import java.io.IOException


/**
 * This is not meant to be a full set of tests. For simplicity, most of your samples do not
 * include tests. However, when building the Room, it is helpful to make sure it works before
 * adding the UI.
 */

@RunWith(AndroidJUnit4::class)
class MatchDatabaseTest {

    private lateinit var matchDao: MatchDatabaseDao
    private lateinit var db: MatchDatabase

    @Before
    fun createDb() {
        val context = InstrumentationRegistry.getInstrumentation().targetContext
        // Using an in-memory database because the information stored here disappears when the
        // process is killed.
        db = Room.inMemoryDatabaseBuilder(context, MatchDatabase::class.java)
            // Allowing main thread queries, just for testing.
            .allowMainThreadQueries()
            .build()
        matchDao = db.matchDao
    }

    @After
    @Throws(IOException::class)
    fun closeDb() {
        db.close()
    }

    @Test
    @Throws(Exception::class)
    fun insertAndGetNight() {
        val match = Match(0, "john", "bob", 1, 2)
        matchDao.insert(match)
        val insertedMatch = matchDao.get(1)
        assertEquals(insertedMatch?.player1SetsWon, 1)
    }
}


